requirements.txt
Flask==3.0.3
Flask_Login==0.6.3
flask_mail==0.10.0
Flask_Migrate==4.0.7
flask_sqlalchemy==3.1.1
flask_wtf==1.2.1
Pillow==10.3.0
pytz==2024.1
SQLAlchemy==2.0.29
Werkzeug==3.0.2
WTForms==3.1.2
boto3==1.37.28
python-dotenv==1.1.0
email_validator==2.2.0
scikit-learn==1.6.1
opencv-python-headless==4.11.0.86
apscheduler==3.11.0
pymysql==1.1.1
trimesh==4.6.8
pymeshlab==2023.12